requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.0
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.6
razorpay==1.3.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.14
whitenoise==6.2.0
